[refactor]: 	- split all views into style, logical and template files 	- redirect to Home route from any route typed with url 	- move down tag canvas from App component to WelcomePage component 	- move particular styles to according files
