+++ src/App.vue
@@ -1,12 +1,11 @@
 <template>
   <div id="app">
     <Root />
-    <canvas class="model"></canvas>
   </div>
 </template>
 
 <script>
-import Root from './views/Root.vue'
+import Root from './views/Root/Root.vue'
 
 export default {
   name: 'App',
