+++ src/assets/styles/index.sass
@@ -20,90 +20,3 @@ body
     min-height: 100vh
     padding: 0
     margin: 0
-
-.main_background,
-.loading_background
-    box-sizing: inherit
-    position: absolute
-    top: 0
-    left: 0
-    width: 100%
-    height: 100%
-    padding: 0
-    margin: 0
-    background-image: url('../../assets/imgs/space-md.jpg')
-    @media (min-width: 720px)
-        background-image: url('../../assets/imgs/space-hr.jpg')
-    background-repeat: no-repeat
-    background-position: center
-    background-size: cover
-    overflow: hidden
-    transition: top 1.3s ease-out
-
-.loading_background
-    z-index: 100
-    opacity: 1
-    filter: blur(12px)
-    transition: opacity 3s ease-out, filter 4s ease-out, z-index 1s linear 1s
-
-.loading_background.loaded
-    z-index: -100
-    filter: none
-    opacity: 0
-
-.main_background.invisible
-    transition: top 1.3s ease-in
-    top: -100%
-
-canvas.model
-    position: absolute
-    display: block
-    top: 50%
-    left: 50%
-    transform: translate(-50%, -50%)
-    width: 95%
-    transition: top .8s ease-out .3s
-
-canvas.model.invisible
-    transition: top .8s ease-in .3s
-    top: -100%
-
-.link-home.hidden
-    top: -5rem
-.hidden
-    opacity: 0
-    transition: all 1.1s ease-out
-
-.link-home,
-.link-planets
-    display: none
-
-.link-home.loaded,
-.link-planets.loaded
-    display: block
-    z-index: 10
-    position: absolute
-    text-transform: uppercase
-    font-weight: bold
-    font-size: 2rem
-    @media (min-width: 720px)
-        font-size: 2.5rem
-    color: white
-    top: 43%
-    left: 50%
-    transform: translate(-50%, -50%)
-    transition: all 1.1s ease-out
-    list-style-type: none
-    transition: all 1.1s ease-out .5s
-    > a
-        text-decoration: none
-        color: white
-    > p
-        font-size: 1.5rem
-        @media (min-width: 720px)
-            font-size: 2rem
-
-.link-home.loaded
-    left: 20%
-    font-size: 2rem
-    top: 3rem
