+++ src/components/PlanetsList/index.vue
@@ -18,5 +18,7 @@
     </div>
   </div>
 </template>
-<style src="./style.sass" lang="sass"></style>
+
 <script src="./component.js" />
+
+<style src="./style.sass" lang="sass" scoped></style>
