+++ src/components/WelcomePage/index.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="WelcomePage">
+    <canvas class="model"></canvas>
   </div>
 </template>
 <style src="./style.sass" lang="sass" scoped></style>
