+++ src/components/WelcomePage/style.sass
@@ -1,4 +1,16 @@
-.WelcomePage
+canvas.model
+    position: absolute
+    display: block
+    top: 50%
+    left: 50%
+    transform: translate(-50%, -50%)
+    width: 95%
+    transition: top .8s ease-out .3s
+
+canvas.model.invisible
+    transition: top .8s ease-in .3s
+    top: -100%
+
