+++ src/main.js
@@ -19,3 +19,5 @@ new Vue({
   router,
   render: h => h(App)
 }).$mount('#app')
+
+router.replace({ path: '/Home', redirect: '/' })
