+++ src/views/Home/Home.vue
@@ -1,0 +1,7 @@
+<template>
+    <WelcomePage />
+</template>
+
+<script src="./component.js"></script>
+
+<style lang="sass" src="./index.sass" scoped></style>
