+++ src/views/Planets/Planets.vue
@@ -1,0 +1,7 @@
+<template>
+    <PlanetsList />
+</template>
+
+<script src="./component.js"></script>
+
+<style lang="sass" src="./index.sass" scoped></style>
