+++ src/views/Root/Root.vue
@@ -1,0 +1,32 @@
+<template>
+  <span class="root_container">
+    <span class="loading_background"></span>
+      <Loading :loading="isLoading" />
+      <span class="main_background"></span>
+      <router-link
+        to="/Home"
+        v-slot="{ href, navigate }"
+        custom
+        >
+        <li class="link-home hidden" @click="toggleVisibility">
+          <a :href="href"  @click="navigate">Home</a>
+        </li>
+      </router-link>
+      <router-link
+        to="/Planets"
+        v-slot="{ href, navigate }"
+        custom
+        >
+        <li class="link-planets" @click="toggleVisibility">
+          <p class="main_paragraph">Star Wars<br>Universum</p>
+          <a :href="href" @click="navigate">Enter</a>
+        </li>
+      </router-link>
+      <router-view/>
+      <Footer :class="!isLoading ? classLoaded : null"/>
+  </span>
+</template>
+
+<script src="./component.js"></script>
+
+<style lang="sass" src="./index.sass" scoped></style>
